fix: properly deserialize the announcement actions
