+++ packages/transport/src/modules/announcements/data/Announcement.ts
@@ -1,5 +1,6 @@
 import { ISerializable, Serializable, serialize, type, validate } from "@js-soft/ts-serval";
 import { CoreAddress, CoreDate, CoreId, ICoreId } from "@nmshd/core-types";
+import { AnnouncementAction, IAnnouncementAction } from "./AnnouncementAction";
 
 export interface IAnnouncement extends ISerializable {
     id: ICoreId;
@@ -10,11 +11,6 @@ export interface IAnnouncement extends ISerializable {
     body: string;
     iqlQuery?: string;
     actions: IAnnouncementAction[];
-}
-
-export interface IAnnouncementAction {
-    displayName: string;
-    link: string;
 }
 
 export enum AnnouncementSeverity {
@@ -54,25 +50,10 @@ export class Announcement extends Serializable implements IAnnouncement {
     public iqlQuery?: string;
 
     @validate()
-    @serialize()
+    @serialize({ type: AnnouncementAction })
     public actions: AnnouncementAction[];
 
     public static from(value: IAnnouncement): Announcement {
-        return this.fromAny(value);
-    }
-}
-
-@type("AnnouncementAction")
-export class AnnouncementAction extends Serializable implements IAnnouncementAction {
-    @validate()
-    @serialize()
-    public displayName: string;
-
-    @validate()
-    @serialize()
-    public link: string;
-
-    public static from(value: IAnnouncementAction): AnnouncementAction {
         return this.fromAny(value);
     }
 }
